mkn_attendance/requirements.txt
asgiref==3.5.2
Django==4.1.1
django-filter==22.1
djangorestframework==3.13.1
gunicorn==20.1.0
PyJWT==2.4.0
pytz==2022.2.1
sqlparse==0.4.2
tzdata==2022.2
